Flair submissions with # of deltas
I added the ability to flair posts with the number of deltas in the comments. The new lines of code are commented out.

The catch is that it doesn't count old comments before the code started running; it only counts the number of deltas accumulated since the new code started running.